refactor(test): collapse e2e family to one parameterized suite; split sqlite repos
- roleplay-behaviors.e2e.test.ts replaces cold-start/exclusion-cascade/
  no-op-inhibition/private-channel-motive (4 suites, 38 its -> 23 its)
- cross-cutting invariants stay owned by pipeline-invariants (no re-assertion)
- sqlite-repositories.test.ts (47 its, 6 suites + contract runs) split into
  6 per-repo files + shared sqlite-test-helpers.ts; 81 tests preserved
- suite: 65 files/685 its/728 tests -> 67 files/664 its/713 tests;
  audit flags 20 -> 6
