+++ docs/test-inventory.json
@@ -1,8 +1,8 @@
 {
-  "generatedAt": "2026-08-19T03:10:23.171Z",
+  "generatedAt": "2026-08-19T03:17:42.143Z",
   "totals": {
-    "files": 65,
-    "its": 685
+    "files": 67,
+    "its": 664
   },
   "byPackage": {
     "engine": {
@@ -19,11 +19,10 @@
       "flags": []
     },
     "server": {
-      "files": 43,
-      "its": 376,
+      "files": 45,
+      "its": 355,
       "flags": [
         "console(1)",
-        "over-cap(47>25)",
         "weak-terminal-toBeDefined(697,775)",
         "async-pause(1)"
       ]
@@ -403,48 +402,6 @@
       "flags": []
     },
     {
-      "file": "packages/server/src/__e2e__/cold-start.e2e.test.ts",
-      "pkg": "server",
-      "layer": "e2e",
-      "its": 9,
-      "itsEach": 0,
-      "forcedCasts": 0,
-      "tsBypasses": 0,
-      "eslintDisables": 0,
-      "onlys": 0,
-      "skips": 0,
-      "consoleLogs": 0,
-      "truthy": 0,
-      "defined": 1,
-      "deadIdTruthy": 0,
-      "nonNull": 0,
-      "spies": 0,
-      "asyncPauses": 0,
-      "weakTerminals": [],
-      "flags": []
-    },
-    {
-      "file": "packages/server/src/__e2e__/exclusion-cascade.e2e.test.ts",
-      "pkg": "server",
-      "layer": "e2e",
-      "its": 9,
-      "itsEach": 0,
-      "forcedCasts": 0,
-      "tsBypasses": 0,
-      "eslintDisables": 0,
-      "onlys": 0,
-      "skips": 0,
-      "consoleLogs": 0,
-      "truthy": 0,
-      "defined": 1,
-      "deadIdTruthy": 0,
-      "nonNull": 0,
-      "spies": 0,
-      "asyncPauses": 0,
-      "weakTerminals": [],
-      "flags": []
-    },
-    {
       "file": "packages/server/src/__e2e__/html-snapshot.e2e.test.ts",
       "pkg": "server",
       "layer": "e2e",
@@ -468,27 +425,6 @@
       ]
     },
     {
-      "file": "packages/server/src/__e2e__/no-op-inhibition.e2e.test.ts",
-      "pkg": "server",
-      "layer": "e2e",
-      "its": 10,
-      "itsEach": 0,
-      "forcedCasts": 0,
-      "tsBypasses": 0,
-      "eslintDisables": 0,
-      "onlys": 0,
-      "skips": 0,
-      "consoleLogs": 0,
-      "truthy": 0,
-      "defined": 2,
-      "deadIdTruthy": 0,
-      "nonNull": 0,
-      "spies": 0,
-      "asyncPauses": 0,
-      "weakTerminals": [],
-      "flags": []
-    },
-    {
       "file": "packages/server/src/__e2e__/pipeline-invariants.e2e.test.ts",
       "pkg": "server",
       "layer": "e2e",
@@ -510,10 +446,10 @@
       "flags": []
     },
     {
-      "file": "packages/server/src/__e2e__/private-channel-motive.e2e.test.ts",
+      "file": "packages/server/src/__e2e__/roleplay-behaviors.e2e.test.ts",
       "pkg": "server",
       "layer": "e2e",
-      "its": 10,
+      "its": 17,
       "itsEach": 0,
       "forcedCasts": 0,
       "tsBypasses": 0,
@@ -522,7 +458,7 @@
       "skips": 0,
       "consoleLogs": 0,
       "truthy": 0,
-      "defined": 1,
+      "defined": 4,
       "deadIdTruthy": 0,
       "nonNull": 0,
       "spies": 0,
@@ -909,10 +845,52 @@
       "flags": []
     },
     {
-      "file": "packages/server/src/persistence/__tests__/sqlite-repositories.test.ts",
+      "file": "packages/server/src/persistence/__tests__/sqlite-agent-state-repository.test.ts",
       "pkg": "server",
       "layer": "integration",
-      "its": 47,
+      "its": 8,
+      "itsEach": 0,
+      "forcedCasts": 0,
+      "tsBypasses": 0,
+      "eslintDisables": 0,
+      "onlys": 0,
+      "skips": 0,
+      "consoleLogs": 0,
+      "truthy": 0,
+      "defined": 0,
+      "deadIdTruthy": 0,
+      "nonNull": 0,
+      "spies": 0,
+      "asyncPauses": 0,
+      "weakTerminals": [],
+      "flags": []
+    },
+    {
+      "file": "packages/server/src/persistence/__tests__/sqlite-channel-repository.test.ts",
+      "pkg": "server",
+      "layer": "integration",
+      "its": 9,
+      "itsEach": 0,
+      "forcedCasts": 0,
+      "tsBypasses": 0,
+      "eslintDisables": 0,
+      "onlys": 0,
+      "skips": 0,
+      "consoleLogs": 0,
+      "truthy": 0,
+      "defined": 0,
+      "deadIdTruthy": 0,
+      "nonNull": 0,
+      "spies": 0,
+      "asyncPauses": 0,
+      "weakTerminals": [],
+      "flags": []
+    },
+    {
+      "file": "packages/server/src/persistence/__tests__/sqlite-event-repository.test.ts",
+      "pkg": "server",
+      "layer": "integration",
+      "its": 13,
       "itsEach": 0,
       "forcedCasts": 0,
       "tsBypasses": 0,
@@ -927,9 +905,70 @@
       "spies": 0,
       "asyncPauses": 0,
       "weakTerminals": [],
-      "flags": [
-        "over-cap(47>25)"
-      ]
+      "flags": []
+    },
+    {
+      "file": "packages/server/src/persistence/__tests__/sqlite-foreign-key-cascade.test.ts",
+      "pkg": "server",
+      "layer": "integration",
+      "its": 2,
+      "itsEach": 0,
+      "forcedCasts": 0,
+      "tsBypasses": 0,
+      "eslintDisables": 0,
+      "onlys": 0,
+      "skips": 0,
+      "consoleLogs": 0,
+      "truthy": 0,
+      "defined": 0,
+      "deadIdTruthy": 0,
+      "nonNull": 0,
+      "spies": 0,
+      "asyncPauses": 0,
+      "weakTerminals": [],
+      "flags": []
+    },
+    {
+      "file": "packages/server/src/persistence/__tests__/sqlite-memory-repository.test.ts",
+      "pkg": "server",
+      "layer": "integration",
+      "its": 8,
+      "itsEach": 0,
+      "forcedCasts": 0,
+      "tsBypasses": 0,
+      "eslintDisables": 0,
+      "onlys": 0,
+      "skips": 0,
+      "consoleLogs": 0,
+      "truthy": 0,
+      "defined": 0,
+      "deadIdTruthy": 0,
+      "nonNull": 0,
+      "spies": 0,
+      "asyncPauses": 0,
+      "weakTerminals": [],
+      "flags": []
+    },
+    {
+      "file": "packages/server/src/persistence/__tests__/sqlite-simulation-repository.test.ts",
+      "pkg": "server",
+      "layer": "integration",
+      "its": 7,
+      "itsEach": 0,
+      "forcedCasts": 0,
+      "tsBypasses": 0,
+      "eslintDisables": 0,
+      "onlys": 0,
+      "skips": 0,
+      "consoleLogs": 0,
+      "truthy": 0,
+      "defined": 0,
+      "deadIdTruthy": 0,
+      "nonNull": 0,
+      "spies": 0,
+      "asyncPauses": 0,
+      "weakTerminals": [],
+      "flags": []
     },
     {
       "file": "packages/server/src/simulation/__tests__/channel-registry.test.ts",
